extract BaseEpisode Feature
