+++ Libs/RickAndMortyEpisodesLib/Sources/Features/EpisodeBrief.feature.swift
@@ -5,19 +5,6 @@ import SwiftUI
 
 /// Namespace for the EpisodeBrief feature. Serves as an anchor for project navigation.
 enum EpisodeBriefFeature {
-  static func formatAirDate(episode: EpisodeDomainModel) -> String {
-    let inputDateFormatter = DateFormatter()
-    inputDateFormatter.dateFormat = "MMMM dd, yyyy"
-    let outputDateFormatter = DateFormatter()
-    outputDateFormatter.dateFormat = "dd/MM/yyyy"
-
-    let inputValue = episode.airDate
-    guard let inputDate = inputDateFormatter.date(from: inputValue)
-    else { return inputValue }
-    let outputString = outputDateFormatter.string(from: inputDate)
-    return outputString
-  }
-
   struct FeatureView: View {
     var episode: EpisodeDomainModel
 
@@ -36,7 +23,7 @@ enum EpisodeBriefFeature {
           Text("\(episode.episode)")
             .font(.caption)
             .fontDesign(.monospaced)
-          Text("\(EpisodeBriefFeature.formatAirDate(episode: episode))")
+          Text("\(BaseEpisodeFeature.formatAirDate(episode: episode))")
             .font(.caption)
             .fontDesign(.monospaced)
           Spacer()
@@ -47,9 +34,5 @@ enum EpisodeBriefFeature {
 }
 
 #Preview {
-  let episode = try! Transformers.loadFixture(
-    output: EpisodeDomainModel.self,
-    fixtureName: "episode_pilot"
-  )
-  EpisodeBriefFeature.FeatureView(episode: episode)
+  EpisodeBriefFeature.FeatureView(episode: BaseEpisodeFeature.previewEpisode())
 }
